debt-collectibility/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
tenacity>=8.2.3
phonenumbers>=8.13.40
rapidfuzz>=3.9.6
pydantic>=2.8.2

# Dev (optional)
pytest>=8.3.2
ruff>=0.5.6

